+++ docs/reference/aligners/punctuation_forced_aligner.md
@@ -1,0 +1,3 @@
+# Punctuation Forced Aligner
+
+::: speechline.aligners.PunctuationForcedAligner
