+++ docs/reference/metrics/phoneme_error_rate.md
@@ -1,0 +1,3 @@
+# Phoneme Error Rate
+
+::: speechline.metrics.PhonemeErrorRate
