+++ docs/reference/pipelines/audio_classification_with_padding.md
@@ -1,3 +1,3 @@
-# Audio Classification with Padding Pipeline
+# Audio Classification with Padding
 
 ::: speechline.pipelines.AudioClassificationWithPaddingPipeline
