--- docs/reference/utils/aligner.md
@@ -1,3 +1,0 @@
-# Punctuation Forced Aligner
-
-::: speechline.utils.aligners.PunctuationForcedAligner
