--- docs/reference/utils/metric.md
@@ -1,3 +1,0 @@
-# Phoneme Error Rate
-
-::: speechline.utils.metrics.PhonemeErrorRate
