+++ speechline/metrics/__init__.py
@@ -12,4 +12,6 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-__version__ = "0.0.1"
+from .phoneme_error_rate import PhonemeErrorRate
+
+__all__ = ["PhonemeErrorRate"]
