+++ speechline/modules/__init__.py
@@ -12,8 +12,8 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-from .audio_module import AudioModule
 from .audio_classifier import AudioClassifier
+from .audio_module import AudioModule
 from .audio_transcriber import AudioTranscriber
 
 __all__ = ["AudioModule", "AudioClassifier", "AudioTranscriber"]
